Change the set of calls in the NEW section.
We have introduced a IS_READ flag in the call log that is used to
determine which entries (specifically voicemail) have not been marked as
read by the user.

Items with IS_READ set to 0 should be placed in the "new section" of the
call log, together with those calls which have the NEW flag set, i.e.,
which are unacknowledge by the user. Unfortunately the terminalogy is a
bit misleading since the NEW field is an existing field in the call log
content provider and its name cannot be changed.

This commit updates both the query used to generate the list of items in
the "new section" and "older section," as well as the query used to
reset the new calls flag. In the former case, we used to reset the NEW
flag only for missed calls, while now we reset it for all types of
calls, making the treatment of the NEW field overall more consistent.

Bug: 5036195
Change-Id: Ib777179a623b5cc97a1f7864cdaca8b5d6b468e9