Use group summary URI
- This is part 1 of showing more details in the group list.
This CL just uses an existing group URI and sets up the
UI to read from a cursor as the user scrolls the list
(instead of reading the entire list and storing it in memory)
- Include count of # members in a group as a subtitle for
each group
- Include the header as an optionally visible part of the
list item because we don't know how many groups are in an account.
Only show the header if it's the first group in an account.
- Refactor some of the code to make it simpler

Bug: 5039532
Change-Id: Ie01fd005416992c4a1b4d89a1591942dc601c894