Fix photo + basic contact info header on tablet
- Fix these cases:
   * On tablet landscape, for a contact with social updates, the name
has to be displayed to the right of the photo.
   * On tablet landscape, for a contact without social updates, the
photo must be static and the name should scroll with the list of
details

- In order to support these cases, specify the header of the list
in XML in the case with and without social updates, otherwise there will
be a lot of boolean flags and unnecessary views inflated

- There still needs to be a flag in the tablet landscape view
because the ContactDetailFragment is reused for both cases where the
contact does and does not have social updates. The static photo
must be explicitly hidden in the case of the contact with social
updates because it will appear in the scrolling header. This flag
must be set outside the ContactDetailFragment b/t the phone
landscape view has a static photo that is shown regardless
of whether the contact has social updates or not.

Bug: 4689488

Change-Id: Ifea16c48496b8552b313a05cbb8ade80ae117069