Add actions to update items.
This commit let you click on an update item if it has an associated
action and actionUri.

In order to make this visible to the user, it adds a background that
will be highlighted when the item is selected. It also marks the item as
focusable for accessibility.

In order to make the selection expand to the entire row, play around
with the padding a little bit: basically, the update list itself is
now as wide as the screen, and the individual items have padding as
appropriate.

Bug: 5095755
Change-Id: Ib1b2d179152beae125dded1b393b3dfc8b22abc9