Show all contacts on invalid contact selection
Happens when a contact is created but it isn't shown due to
contact fliter (e.g. the contact is for the account A while the
filter shows only contacts for the account B). Single contact
mode is just confusing in that case, so now let's show all contacts
and notify the change to the user.

Bug: 5031526
Change-Id: Ica75178d10b4a9b95c44cc4f75f56b65fbc67da8