Call log auto refresh.
Modified ExtendedCursor to support registration/unregistration
of contentObserver & datasetObserver on the cursor.

Automatic refresh of voicemail when inserted through the voicemail
content provider still does not work because call log uri is not
notified in this case. This will require a fix in the content provider.

Moved the call to reset new flag from onResume() to onStop() such that
calls are now marked as read when the user leaves the call log screen.
If entries were continued to be marked as read in onResume() then it
triggers a refresh and moves the entries from "new" to the
"old" section immediately. We want the entrie to remain in the "new"
section until the user exits the call log screen.

bug: 5055868

Change-Id: Iaef05ae721df1ab19dc001e17f4cd7be4019863e