Clear missed call notification.
The current implementation failed to clear missed call notification
because it requires the parent activity's window to in focus.

Due to some reasons getActivity().hasWindowFocus() returns false when
invoked from onResume(). By experimenting it came up that the window
gets focus AFTER the fragment is resumed.

The fix is to to use KeygaurdManager.inKeyguardRestrictedInputMode() to
determine if the keygaurd is on, and not remove the notification if so.

Bug: 4521535

Change-Id: I56cf82d708d7c694a966bdd6b79a328895c86048