Support for application-provided images in stream items.
This change add a way to include an image in the text and comments of a
stream item.

The image can be included by having an <img> tag whose src attribute is
a URI identifyin of resource to use of the form:

    res://package/name

where package is the name of the package containing the resource and
name is the name of the resource as defined by the package (notice, not
the id, but the name of the resource).

Bug: 5123531
Change-Id: I8f35fd1f98a2b9cc8f5223d67f8c143e2714716e