Only call onVisibilityChanged on creation for the current tab.
When the fragment is initially attached, there is a call to
onVisibilityChanged, so that the fragment knows whether it is shown or
not.

Currently, we are calling this with true or false for all tabs (actually
only the DialpadFragment and CallLogFragment, as they implement the
listener interface). However, we do not need to be notified that we are
not visible if we are just being created.

This is needed because otherwise we will assume we were being dismissed
while in reality we were never shown.

Bug: 5157085
Change-Id: I37be7799b0e473cc2eabefef75a56335420979a3