Make tests more resiliant to changes.
Instead having a number of tests create the values for the call log
cursor, create them always using a helper class, so that the tests will
always be in sync when the code is changed.

Since I plan to add a few more columns to the call log query, this will
save some time later on.

Bug: 5101753
Change-Id: I366de7502a03011dbff8f7659d907e89c389c382