Do not show geocode for voicemail number.
As I moved the geocoding to the content provider, we lost the fact that
the voicemail number should not be geocoded, because the voicemail
number can change over time and the provider does not know what it is.

Instead, this change adds an explicit check in the UI code.

Note that this was caught by a unit test, but unfortunately I did not
notice I broke it when I moved the geocoding to the content provider.

Bug: 5111400
Change-Id: Ia14d8ac1a22b52e36bdac784fad5fea706865cb7