Fix com.android.contacts.detail.StreamItemAdapterTest
In a recent change, I added a header in the stream items adapter.
This means all the indices must be incremented by one in the tests, to
match the new layout: the adapter now contains always a header, hidden
under the tab carousel (if present), a title, and the stream items.

Bug: 5162546
Change-Id: I4e38b1cf513e316a3a42fb22094b0b4b16f76a69