Use the call log to cache additional contact information.
The new version of the call log needs a bit more information about the
matching contact (if any) to render its entries.

This change both uses the newly added columns in the call log to lookup
this values and stores the updated values back into the call log.

These values are used for setting up the quick contact badge and
formatting the phone number.

This also fixes a few smaller issues:
- the phone number shown used to change format between the initially set
  value to the one obtained from the contact: this value is now cached.
- we were using -1 as the id of a non-existing contact, but 0 is
  actually the value used elsewhere for this, so stick to that.
- make sure we use the cached value from the call log for formatting.
- update the call log if there was no previous value.
- introduce a UriUtils with a couple of useful functions for handling
  URIs.

Bug: 5101753
Change-Id: I22f2b03b234da91e02dbeebe11a94c2f56295a3d