Fix group source crash on phone
- The group source button is supposed to be a scrolling header
in the list of group members on the phone portrait view, but
we can't add a header view to a ListView after the adapter is set.
(Setting the adapter had to be moved before setting the header
view otherwise the group detail page would flicker on every
sync).

- Make the header static as a simplification and hide the view
if there's no group source needed.

- Remove fading edge on group members list

Bug: 5195472
Change-Id: Ibedc629ddd85a54ed8beee92650df617a09d2a6c