Highlight the current group on group list
Because the group list item may contain the header view, we can't just
set the background to the list item.
So tweaked the layout a bit and set the background to the actual body
part, using duplicateParentState.

Also disabled the default listview selector and added the pressed state
for the background drawable, so now the pressed state works file.
(Except the drawable we use is different from the framework default one --
I'm asking for new assets.)

I added a new selector drawable group_list_item_background.xml rather
than reusing the one used for the contact list, because what we want
is actually different; we don't want to disable the default listview selector
for the contact list, but we only do this for the group list.

Bug 5186255

Change-Id: I5f0a6488b7352368b903b04e28f51905374f54d8