Persist selected group on rotation
It was all about timing of restoration.  We did have the code to restore it,
but it should have been done before mAdapter.setSelectedGroup().

Also switched from Bundle.putString() to putParcelable().
(putParcelable accepts null -- so the change should be safe.)

(This CL also restores the activated item on the group browse list, although
we don't set proper background to it.)

Bug 5137176

Change-Id: I735f80419c60bb3dd5e760c639900b4eba1ef6c5