Let Phone UI reload filter on onStart()
This is workaround for bug 5165507: Contacts to Display filter loses
its setting when going from People to Phone app

- call ContactListFilterController#onStart() or its variant.
  Not calling it is an apparent bug in DialpadActivity
- instead of checking local filter cache, load it from preferences
  every time, only on Phone UI

The root problem is that People and Phone UI don't share one
controller insntance and thus they don't receive the same
filter change event. When the user changes the filter in People,
Phone UI won't receieve the event, for example.

On Phone UI, back button doesn't introduce Activity#finish(), which
makes the whole issue more prominent in People -> Phone case
(set filter on People, and go to Phone), but it does exist in the
opposite case (set filter on Phone, and go to People). The latter
case causes the problem only when People is in background (with
Home button, for example).

This will fix:
- display filter change isn't reflected in Phone, when going from
  People to Phone (what reported in bug 5165507)

This won't fix:
- display filter change isn't reflected when it is modified in
  Phone UI, while People is in background (with Home button event)
  (what caused by the root problem behind that bug)

A more fundementaly fix would be to share one single
ContactListFilterController, which would be more complicated and
a bit too risky at this point.

Bug: 5165507
Change-Id: Ie7968deb99c07597c35e3bb1531461f0a42c326a