Fix blank screen on VIEW raw contact URIs or legacy URIs on phone
Store requested URI in ContactLoader Result. We were doing the
wrong comparison of the lookup URI. For example, we were checking
that the incoming URI from the intent (i.e. raw contact URI) was
equal to the lookup URI from the ContactLoader Result. This was
not equivalent and resulting in a blank screen for the VIEW
intent on raw contact URIs and legacy URIs.

TODO: Fix for tablet, there may be a different issue going on
there.

Bug: 5220640
Change-Id: I2d6dff8df4c9065d7d845dca5c09ba1d773770df