Prevent attribution/ comments from getting cut off
This is the new design for bug:5236510 where text is shown
below (instead of beside) an image.

Makes sure that the attribution and comments lines always show
and that the content does not get cut off in the vertical middle
of a line.

Remove quotes from updates tab text.

Add margin on the tablet portrait contact card view.

Change-Id: Ica021a807a7ab1178e305d3eab6347133ffc3595