Choose ALL filter when there's no filter saved yet
We currently choose CUSTOM filter when there's no filter saved
in shared preferences. It was for keeping users' choice in GB.

Note that in GB we didn't have filter mechanism at all, and what we
had at that moment is what we call "custom filter" in ICS
(custom filter == showing contacts in IN_VISIBLE_GROUP on our
ContactsProvider2)

The problem is that, the fallback choice is not only used after
GB-to-ICS migration but also used when users boot the device
for the first time (after purchasing it, for example). Showing
"CUSTOM FILTER" for the first time of using Contacts app looks really
weird. Also we don't have no easy way to determine if the first app
launch is from migration or not.

Bug: 5241896
Change-Id: Ib151518837b063cf5071a924919d492385885b76