Re-introduce appropriate "activated" background
On tablet, we need "triangle" on the right side of the asset.
I7c08d8f533f0dd5318c614e3df14bb39207485e9 removed it.

This change creates the directories for tablets instead of
replacing existing ones. We may have activated state on
phone, which don't want the custom asset.

Bug: 5248893
Change-Id: I9486a20991aae2c66a7216fc9e854a12c9e8f030