Use standardized no account screen on first launch
- If there are no accounts on the device, then launch
an intent to show the prompt (check that there is no
flag in SharedPreferences preventing the prompt from
being shown).

- Only when the user explicitly selects "Not now" from
the no account screen, do we store a flag that says
we should never show the prompt again.

- Add "Manage accounts" permission in order to
perform AccountManager.addAccount, which gives us access
to the "no account" screen.

Bug: 5250360
Change-Id: I2f925c838bc0e7003a8dbb3280e5a22ed8680670