Make sure only People and Dialtacts responds to search key
- Remove the default search activity meta-data from manifest, so the old-style
search dialog won't show up when the serach key is pressed.  It was for
pre-HC style search UI.  Now the search key only works on people and dialtacts,
which implement either startSearch or onSearchRequested.

- Removed all the startSearch/onSearchRequested methods which are now
unnecessary.

- Also removed startSearch() in some fragmets.  They're not used.

- Tested some viewer/editor/picker activities, and also both dialtacts and
people, the launcher search box, and the SEARCH intent, on both the phone
and the tablet (except dialtacts on tablet).  All looking good.

Bug 5230790

Change-Id: If269df74ae218707daa5b7039439c6fea2575e8d