Invalidate options menu after loading the details.
We do not know until we have loaded the details of a call whether we
should allow dialing the number: the call might be from a private or
unknown number. However, the content of the menu is loaded when first
starting the activity, and, by default, does not include the option.

Adding a call to invalidateOptionsMenu() after loading the details
forces the menu content to be updated.

Bug: 5257543
Change-Id: I0581c278da74ff2dd252e066fb4fbfdb3759dbb0