Maintain the current view (voicemail only or all calls) on updates.
When the underlying content provider content changes, keep the same view
(voicemail only or all calls) as the one the user is looking at.

The underlying data may change because we updated some of the cached
columns, which currently causes the view to be updated for what seems to
the user no valid reason.

The only events we want to be sure to communicate to the user are new
voicemail and missed calls, but the notification bar takes care of those
cases, and if the application is currently open, we will refresh the
content anyway when resumed.

Bug: 5247042
Change-Id: I6db0a980aaafed5990f8e84bf4a9a1d0f44efb36