Use OpenableColumns.DISPLAY_NAME for import message
Previously we've used the last path segment of a given URI,
which may not be correct. DISPLAY_NAME will be more appropriate
especially when vCard view intent is sent from the other apps.

Note that a display name can be available only when the external
apps support OPENABLE category, while we are using DEFAULT category
for VIEW vCard intent. In the future we should change the category
too, but for now this change leaves it as is.

TESTED:
- import from SDCard
- import from Email apps
- import via BT contact share

Bug: 5287116
Change-Id: Iad572eb48b6756246b15d992a8d4231a1b4cb31d