Extend the call log highlight to include the call icon.
This commit moves the call icon (the secondary action) within the
primary action, which means the highlight for the primary action will
extend to it.

In order to preserve accessibility, it add a nextFocusRight and
nextFocusLeft to the primary and secondary actions, so that it is
possible to navigate between them.

Bug: 5290460
Change-Id: I75f1ca5530d9e656d54f32b58bc1cb04bfc6bed6