Make quick contact badge accessible.
Since the quick contact badge is now within the primary action, we need
a nextFocusLeft to be able to select it.

Note that it was not selectable earlier because it was not set as
focusable. Also, currently the focused state is not visible: it would be
nice to fix that as well.

Bug: 5290460
Change-Id: Ib2a2ce9ee42daef54217df66e2faaf7d11cff570