Sets the minimum height on the ActionsViewContainer.
Since the frame layout containing the actions had a minimum height but
not the ActionsViewContainer itself, it was possible to have extra space
around it which would not be highlighted correctly.

Moving the minimum height to the ActionsViewContainer itself means that
there will never be extra padding from the FrameLayout.

Also remove the declared top and bottom padding since these are always
set programmatically anyway.

Bug: 5341956
Change-Id: I0345e1bf6f27cb92c499adf0b1e9ce1363155058