Improve D-Pad accessiblity of the call details.
We are doing something fancy with the layout, to scroll the header when
the list view is scrolled: a list view is shown under the controls, with
a first header item which is invisible as it is behind the controls; the
controls are then scrolled up when the list is scrolled. This is similar
to what is done in the tab carousel in the contact details/updates.

Unfortunately, this broke the accessibility of the call details.

In this commit I have a partial fix: whenever the hidden header is
focused, focus is transfered to the controls. This makes the history
items not accessible, but makes the more crucial controls accessible
again.

Bug: 5323306
Change-Id: I342e58b9052e2aba3d5d8c8ea1b178108741de91