Fix name editor data loss on rotation
Bug: 5290661 Edited name for a contact not saved

When the contact editor was recreated on screen rotation the
data on the name editor was updated incorrectly.
The problem was that a change to any of the fields in the name editor
was erasing the other fields. Instead of erasing the other fields, i changed
the code to sync between the display name and structured name so that on
rotation, the data is correctly stored and retrieved.

Change-Id: I644bfea2af4e759e1d7d94aa09840ed15efd079d