Avoid showing error if contact removed on save.
Because removing all data from a local contact can actually cause
the raw contact and contact to be deleted, we need to track that
case so that the caller does not consider the lack of a lookup
URI to be an error.

Also made "Me" header all-caps for no-profile case.

Bug 5346614
Bug 5354364

Change-Id: If1e1d67da9c14eb8782be05b2e39ece19b5ac026