SelectAccountDialogFragment: Handle orientation changes
- Put all parameters to the argument Bundle, so they won't be lost
on orientation changes.

- Removed the subdialog in ImportExportDialogFragment, which used
the argument Bundle for a different purpose.  Now we put the
information needed to ImportExportDialogFragment as a sub Bundle
in the argument Bundle.

Tested all 3 uses:
- "Add new group"
- "Copy to my contacts" (Couldn't configure my device to show this dialog,
so I had to apply a local patch to test this part.)
- "Import from SD" "Import from SIM".
Also made sure other features on "Export/Import" still work, except for
pre-existing bug 5387545.

Bug 5382646
Bug 5369853

Change-Id: If7f75ee8192889dbe484874c52e1a355d6b35c6c