Remove strict mode violation with number formatting.
The call log is currently performing formatting on phone numbers in the
main thread (during bind view).

At the same time, however, we are always doing a background request to
look-up the contacts. Move the formatting call to the background thread.

In order for the right information to be shown as soon as the call log
is opened, we also need to store this information in our cached values
in the database. This means that the number will show unformatted the
very first time, and then updated once the background require is
completed.

Bug: 5316982
Change-Id: If1e4990243e73af6c9519fbf5e76a5d234e99e15