Fix updating of seek bar position while seeking.
We stop updating the seek bar position in the background while seeking.
However, if the background task was running at the time we stop it, it
will still update the position once more, with the default value of
zero, which made the seek bar jump to zero at what seemed like random
times.

Instead, if the updater has been stopped, simply return immediately.

Bug: 5094799
Change-Id: I3f9f52ffcce442bb19f1f1a6731161d7fc2e94cd