Do not show black header if there is no action.
In the call details activity we show a black overlay on top of the
picture for adding a contact or accessing the contact. For some numbers,
no action is possible, e.g., SIP number without a contact, voicemail
number, private numbers.

In those cases, do not show the black header at all, since there is no
action this header would be describing.

This requires a bit of changes to the scrolling: if the header is not
present, we should scroll out the entire picture and the separator line
out of it.

Bug: 5400504
Change-Id: I2effa6664df6eec6051b5c685cea527f56e049f1